// src/app/modules/admin/stock/models/madeProduct.ts
export interface MadeProduct {
  id: number;
  length: number;
  datePicked: Date;
  status: string;
  productName: string;
  frameType: string;
  stickType: string;
  color: string;
  gauge: number;
  steelCoilId: string;
  steelCoilNumber: string;
  manufacturedDate: Date;
  productTransactionId: number;
}